A0-1-6: Handle alias templates
alias templates are not properly handled by the extractor, so we provide
special case handling that ensures alias template instantiations are
correctly associated with alias templates.

alias templates are not represented using the same template
infrastructure as other templates. Instead, each instantiation is a copy
of the template type, with no specified location. We use this to
identify templates with at least one instantiation. We assume an
instantiated template is used, as the only way it would be instantiated
is if it was used.