Add rule category to rule meta data
The category information enables evaluation time computation of the
effective category of a rule. The effective category is determined by
the rule's category and a possible recategorization through a Guideline
Recategorization Plan.

The effective category will be used to determine how to handle a deviation applied to a query implementing the
rule.