Use the effective category in the exclusion process
In this commit we expand the meta data `Query` class to provide an
effective category that is determined by combining a query's category
and a possible active Guideline Recategorization Plan.

The exclusion process now validates for a query with an associated
deviation record whether the effective category allows a deviation.
If this is not allowed, then the deviation record is not applied.

Any deviation record applied to a query with an effective category not
permitting deviation is considered invalid and will be listed in the
analysis report.