A5-2-2: Exclude c-style casts generated by library macros
casts generated by library macros are not modifiable by the user, so
can be excluded from the results of this rule.

In addition, this commit improves the query message to specify whether
the cast was generated due to the expansion of a macro, and to provide a
link to the macro in question. This helps users identify the location in
the code where the cast is actually written.