EssentialTypes: Add essential types library
This module calculates and reports the essential type of each given
expression in the program, as defined in the MISRA C:2012 standard.

The essential type for an expression is calculated based on the AST
type of the expression. Where it differs from the standard type,
the calculation is overridden to implement the MISRA definition.

Various utility methods related to essential types are included.