M14-6-1: Restrict to templates with dependent base
The query didn't explicitly enforce this condition, and instead relied
on the fact that NameQualifiableElement's without any qualifier would
only point to the "wrong" element within templates. However, there are
some edge cases (static member functions with overrides) where non-class
templates could be flagged.

This commit restricts the output to the set of template classes with
dependent base types, therefore avoiding the sort of false positives
seen above. In addition, this improves performance for this query
because the set of template classes with dependent base types is much
much much smaller than the overall set of base types.