M14-6-1: Handle overloading
Previously the logic of this query asserted that the function call did
not target the selected function in the dependent base type with the
same name. In theory this is wrong, as overloading can permit multiple
functions declarations with the same name, but different signatures, so
we now say that the target is not declared on the same base type.

In practice, such results would be excluded because to call an overload
of the same function would always have a qualifier in this case, however
this logic makes the intention clearer.