M14-6-3: Address perf problems
Performance issues occurred because we are trying to find pairs of
functions or variables which have the same name but are not the same.
The join orderer is very keen on joining two copies of the function/
variable table early on which, on large databases like openpilot with
a lot of name duplication, can cause signficant blow-up.

The workaround is to provide helper predicates that ensure the
restricted set of functions/variables we care about (targets of accesses
or calls in templates with dependent base types) are computed first,
then the name is joined with a member on the dependent base type.