M14-6-1: Identify non-compilable cases
To take a pointer to a member function we need to provide a qualifier to
the name. The cases in the COMPLIANT classes are therefore
non-compilable. The reason some cases were not previously causing EDG to
crash was that the templates weren't instantiated. We now instantiate
all templates, and add explanatory comments.