A5-2-2: Exclude results in uninstantiated templates, explain limitations (#160)
* A5-2-2: Clarify c-style casts scope, exclude templates

Clarify what `CStyleCast` does and does not cover by adding a comment,
expanding the test case and providing an implementation scope. In
addition, exclude casts on template parameters to avoid unnecessary
false positives.

* A5-2-2: Exclude uninstantiated templates

Any cast in an uninstantiated template that is related to the template
parameter may be converted to a `ConstructorCall` when the template is
instantiated. To avoid the common false positive case where the
functional cast notation is used to call a constructor, we exclude all
results in uninstantiated templates and instead rely on reporting
results in template instantiations instead.

---------

Co-authored-by: Mauro Baluda <mbaluda@github.com>