STR37-C: Handle macros in <ctype.h>
The <ctype.h> commonly implements its APIs using either macros or
functions or some combination of the two. Our query only assumed
functions were used, whereas macros are practically used by both gcc and
clang, and these can vary depending on compiler flags.

The CharFunctions.qll library now provides a unified interface from
which to get a unique expression for each use of an API in the library,
hopefully regardless of whether it is a macro or a function. To do this
we have had to hard code assumptions about the structure of the macros,
however our matrix compiler testing should flag if these assumptions are
broken with a particular version of a supported compiler.