FloatingTypes: Implement FLP36-C
Add a query to detect integer to float conversions where precision may
be lost because the mantissa of the float is limited in size.

The strategy here is to assume IEEE754 and standard mantissa size, then
use range analysis to determine an upper bound on the value converted
and check if that exceeds the "safe" upper limit where all below can be
fully represented.