EssentialTypes: Improve performance of Rule 10.5
This improves the performance of Rule 10.5 by computing all the types of
the cast first, before determining whether they were invalid. Previously
the join orderer found the essential type category of the "from" type,
then joined that will invalid type categories, then joined that with all
expressions of that type before filtering down to the cast.