Generate dummy `Query` for shared query test files
Shared query test files were not excluding elements outside the source
archive because they did not implement the `..SharedQuery` mechanism.
This commit adjusts the `Query` class to provide a dummy query for this
purpose, and updates the generator to create an instance of that
abstract class for each shared query.