A7-1-5: Exclude compiler generated variables
A7-1-5 should only apply to non-compiler generated variables.

This commit adds a new test case which shows an example which, on some
platforms, will cause the compiler to generate a __range variable that
is flagged by the query before this change. Note: I have been unable to
replicate the problem in a unit test.