Address string related compiler compatibility issues.
This commit restructures the string, string.h and cstring headers to match the
expected headers and namespaces of C++ standard declarations. This addresses
compatibility issues with real compilers because our test cases expected to
access C string functionality through the "#include <string>" header include,
which is not the case in practice.

This commit addresses compiler compatibility issues in the following rules:
 * A12-0-2
 * A27-0-2
 * M16-0-5
 * M18-0-5
 * DCL55-CPP
 * EXP62-CPP
 * OOP57-CPP
 * STR50-CPP