A15-2-2: Avoid infinite interpretation edge case
In CodeQL CLI 2.12.7 there is a bug which causes an infinite loop
during results interpretation when a result includes more than maxPaths
paths and also includes a path with no edges i.e. where the source and
sink node are the same.

To avoid this edge case, if we report a path where the source and sink
are the same (i.e the throwingExpr directly throws an exception), we
adjust the sink node to report the constructor, which creates a one
step path from the throwingExprFlowNode to the constructor node.

This also means we can delete the `nodes` query predicate, as we only
included it to enable zero-path elements to display.