Suppress dataflow deprecation warning
In CodeQL CLI 2.14 it is not possible to include deprecation warnings in
test files because the paths are absolute, and so are not portable
between different systems. We avoid the deprecation warning for now by
copying the relevant parts of the dataflow library and removing the
deprecation warning.

This workaround can be removed when we upgrade to 2.15, as the
deprecation warning has been fixed to not use absolute paths.