Change definition of out parameter and add exclusions
An out parameter is a non-const reference or pointer that is modified.
We removed assignment operators and crement operators as exclusions to
the modified property, because they should still be used to qualify an
out parameter.

Additionally we exclude user-defined non-member assignment operator and
stream related operator parameters that are required to be passed by
non-const reference.