Exclude flows through non constexpr variables
Before the analysis only considered whether the source of an argument
passed to a function was computed at compile time. Now we consider
whether intermediate variables are also constexpr even though their
values are compile time constants, because otherwise the compiler
will accept the variable receiving the compiled time constant to be a
constexpr variable.