Adjust alert message for A12-4-1 and A12-8-6
For A12-8-6, the alert no longer contains the reason for the base class
being a base class. This is because a base class with both a derived class
and being abstract is reported as a separate alert.

For A12-4-1, the alert contains the fully qualified name of the base class
and is formatted according to the style guide.