Detect compilations with no warnings when '-w' flag is present.
Gcc may be compiled to auto include warnings such as -Wformat. However,
passing in `-w` will suppress the enabled format warnings. The previous
query would not raise an issue, as it saw the `-Wformat` flag etc, even
though if `-w` was present, causing gcc to run with no warnings enabled.