Fix #689, false negatives for A1-1-2 thinking -Wno-foo is compliant.
The presence of -Wno-foo should not mark the compilation compliant with
A1-1-2, nor should the presence of -Wfoo=0.

Easily check for all -Wfoo=bar flags, that foo is not no-baz, and bar is
not 0. Also check there is no -Wno-foo flag overruling it. Otherwise the
query functionality remains the same.

Add test cases for non-compliant scenarios -Wfoo=0 and -Wno-foo, and for
the compliant scenario -Wall -Wno-foo.

This will have some compatibility issues with PR #688, after one is
merged the other will need some small updates before this can be merged.