M0-1-2: Remove reference to GuardCondition
In the upgrade to 2.16.6 GuardCondition has been modified to use
the IR instead of the AST mode it used before. One of the
consequences of this change is that GuardConditions no longer exist
for uninstantiated templates (because the IR does not apply for
uninstantiated templates).

Instead of using GuardCondition we can use the same logic for
deducing infeasible paths for uninstantiated templates as for non
template code. This avoids the dependency on GuardCondition, and
provides consistency with the rest of the query, while shortening
the query.