Use shared queries / generally report obsolescent features even if redundant.
Redundant reports should not be a common user issue; these features are
obsolescent and likely rarely used and less often to be excepted.

Implement ungetc() on a zero-offset stream and specific banning of gets(), as
the redundant rules for those obsolescent features report a far wider set of
issues than banned by RULE-1-5.

Implementation of banning ungetc() on a zero-offset stream is not thorough or
comprehensive. This should be fine. False positives should not create any user
issues because the call of the function overall is banned. And false negatives
should not be an issue, for the same reason.