Pack creation: load help from base reference
For pull requests the current process attempts to load the help
from the equivalent ref in the help repo. As most PRs do not add
an equivalent branch on the help repo, this means that most PRs
do not load any help at all, and the constructed artifacts are
missing help.

This is problematic during the release process, where we appear to
use the artifacts built from the PR event, instead of the artifacts
built from the branch itself, therefore using artifacts without
any help included.

This commit modifies the behaviour to fetch the help for the base
ref for the pull request or merge group. This should ensure that
help files are always loaded, regardless of where the artifacts
are built.