ValidContainerElementAccess: Address new FPs
We had some new false positives because in 2.18.4 string taint is
tracked into the qualifier of a string operation, such as insert.
This caused us to erroneously identify the container itself as a
reference to an element of the container. This has been addressed
by excluding uses of the owning container from pointer or reference
access.