Scope: Remove hides(..) and related predicates
There are no more consumers of hides(..).

In addition, it doesn't make sense conceptually to look for
variables in the same scope with the same name, as scopes will
prohibit using the same name in the same scope. Reviewing real
world cases where this occurs, they all seem to be extractor
oddities (multiple copies of parameters for the same function etc.)
which provides further evidence that this mode is not required.