IdentifierHidden: remove lambda special casing
This removes the special handling of lambda expressions, which
was causing performance issues.

Instead, we rely on the new behviour of the Scope library, which
calculates identifier hiding for lambda expressions as part of the
main calculation.

This has one semantic change - the new code applies
`isInSameTranslationUnit`, which reduces false positives where the
identifier "hiding" in a lambda occurred with an outer variable in
a different translation unit.