A5-1-9: Address param confusion
The hash cons value for parameters was incorrectly calculated with
parameter uses (e.g accesses to the parameter). The correct
approach is to use the variable name and type.

This caused performance issues, because the hash cons for a
function was made up of all combinations of the accesses to the
parameters. For lambdas with many parameters and many accesses,
this was problematic.