Add "Baseline" query suites for C/C++.
These are queries from MISRA, AUTOSAR, and CERT, that have been hand picked
based on a few properties:
- review of violation count for each rule on real projects
- review of accuracy of each rule
- subjective review of broad applicability across most projects
- subjective review of severity of each rule

The baseline rules have been divided into two categories, "style," and "safety."
The choice of which is subjective, but as a general rule:
- Rules that tend to report only erroneous cases are typically "safety."
- Rules that tend to report patterns that obscure erroneous behavior are
  typically "style."
- If alerts can be mitigated by simple changes, it's more likely "style."
- If it is challenging for a human to spot the rule violation, it's more likely
  "safety."
- If the rule is designed to enforce some kind of "good habits," not just fix an
  isolated it's more likely "style."
- If the rule feels "opinionated" in some sense, it's more likely "style."
- If it is difficult to decide whether a finding should be fixed or deviated,
  it is more likely "safety."
- Unsurprisingly, if the rule is purely syntactic or related to naming, it's
  more likely "style."

Some rules are in both suites -- for instance, rules that ban the use of 'goto'
are conceivably both "style" and "safety." While goto is not erroneous in
itself, and gotos are easy to spot during human review, it is typically not a
simple task to refactor code to remove gotos, and it is difficult to decide when
a goto serves a legitimate purpose and the rule should be deviated.

In general, the 'style' suites have more violations which are easier to fix, and
the 'safety' suites have fewer violation that are harder to fix but warrant more
attention to address. Project authors can therefore enable both with the
'default' suite, or pick between the two kinds suitable for their project and
goals.